FAT-1889 BreedPicasso. Добавил метод getTarget
