+++ app/src/main/java/com/andersen/dogsapp/dogs/data/web/BreedPicasso.java
@@ -1,11 +1,20 @@
 package com.andersen.dogsapp.dogs.data.web;
 
 import android.content.Context;
+import android.graphics.Bitmap;
+import android.graphics.drawable.Drawable;
+import android.util.Log;
+import android.view.View;
 import android.widget.ImageView;
+import android.widget.ProgressBar;
 
 import com.andersen.dogsapp.R;
 import com.squareup.picasso.Picasso;
 import com.squareup.picasso.Target;
+
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
 
 public class BreedPicasso {
     private static BreedPicasso instance;
@@ -26,7 +35,7 @@ public class BreedPicasso {
     }
 
     public void initWithTarget(String uriBreedString, Target target){
-        picasso.get()
+                picasso.get()
                .load(uriBreedString)
                .placeholder(placeholder)
                .into(target);
@@ -37,7 +46,45 @@ public class BreedPicasso {
                 .load(uriBreedString)
                 .placeholder(placeholder)
                 .error(R.drawable.smiled_dog_face)
+                .fit()
                 .into(dogKindImageView);
+    }
+
+    public Target getTarget (ProgressBar itemProgressBar, ImageView dogKindImageView, File breedImageFile){
+        return new Target() {
+            @Override
+            public void onBitmapLoaded(final Bitmap bitmap, Picasso.LoadedFrom from) {
+                itemProgressBar.setVisibility(View.INVISIBLE);
+                dogKindImageView.setImageBitmap(bitmap);
+                new Thread(new Runnable() {
+                    @Override
+                    public void run() {
+                        FileOutputStream fos = null;
+                        try {
+                            fos = new FileOutputStream(breedImageFile);
+                            bitmap.compress(Bitmap.CompressFormat.JPEG, 100, fos);
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        } finally {
+                            try {
+                                fos.close();
+                            } catch (IOException e) {
+                                e.printStackTrace();
+                            }
+                        }
+                    }
+                }).start();
+            }
+
+            @Override
+            public void onPrepareLoad(Drawable placeHolderDrawable) {
+                if (placeHolderDrawable != null) {
+                }
+            }
+            @Override
+            public void onBitmapFailed(Exception e, Drawable errorDrawable) {
+            }
+        };
     }
 }
