+++ app/src/main/res/layout/dog_kind_item.xml
@@ -24,7 +24,7 @@
 
             <ImageView
                 android:id="@+id/dog_kind_image_view"
-                android:layout_width="wrap_content"
+                android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_centerInParent="true"/>
